requirements.txt
-i https://pypi.org/simple
asgiref==3.2.10; python_version >= '3.5'
braintree==4.1.0
cachecontrol==0.12.6; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
certifi==2020.4.5.2
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
deprecation==2.1.0
django-extensions==2.2.9
django-render-block==0.6
django-rest-xauth==1.0.6
django-templated-email==2.3.0
django==3.0.7; python_version >= '3.6'
djangorestframework==3.11.0; python_version >= '3.5'
idna==2.9; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
jsonpickle==0.9.6
jwcrypto==0.7
msgpack==1.0.0
packaging==20.4; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pycparser==2.20; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pyparsing==2.4.7; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dateutil==2.8.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytz==2020.1
requests==2.24.0; python_version >= '3.0'
six==1.15.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.3.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
squareup==5.3.0.20200528
stripe==2.48.0
timeago==1.0.14
urllib3==1.25.9; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
